refactor: TableGroupValidator 추가
